Refactor to separate concerns
