+++ frontend/src/views/DeclaredElementPage/index.vue
@@ -6,33 +6,7 @@
       <DsfrAlert v-if="alert" v-bind="alert" class="mb-4" />
       <div v-if="element">
         <div class="grid md:grid-cols-2 gap-4">
-          <div class="bg-grey-975 p-4 mb-8">
-            <p class="text-blue-france-sun-113">
-              <v-icon :name="icon" />
-              {{ typeName }}
-            </p>
-            <div class="grid grid-cols-3">
-              <p :aria-hidden="true" class="fr-h2">{{ authorizationInfo.flag }}</p>
-              <p class="content-center col-span-2">{{ authorizationInfo.text }}</p>
-            </div>
-            <div v-for="(info, idx) in elementProfile" :key="`element-profile-${idx}`" class="grid grid-cols-3">
-              <p>
-                <b>{{ info.label }}</b>
-              </p>
-              <p class="col-span-2">
-                <a v-if="info.href" :href="info.href" target="_blank" rel="noopener" class="text-blue-france-sun-113">
-                  {{ info.text }}
-                </a>
-                <span v-else>{{ info.text }}</span>
-              </p>
-            </div>
-            <div class="grid justify-items-end p-2">
-              <router-link :to="declarationLink" class="text-blue-france-sun-113">
-                Voir la déclaration
-                <v-icon icon="ri-arrow-right-line"></v-icon>
-              </router-link>
-            </div>
-          </div>
+          <ElementInfo :element="element" :type="type" :declarationLink="declarationLink" />
         </div>
         <div>
           <DsfrButtonGroup :buttons="actionButtons" inlineLayoutWhen="md" align="center" class="mb-8" />
@@ -51,69 +25,12 @@
 <script setup>
 import { computed, watch, ref } from "vue"
 import { useFetch } from "@vueuse/core"
-import { getTypeIcon, getTypeInFrench, getApiType } from "@/utils/mappings"
+import { getApiType } from "@/utils/mappings"
 import { handleError } from "@/utils/error-handling"
 import { headers } from "@/utils/data-fetching"
+import ElementInfo from "./ElementInfo"
 
 const props = defineProps({ type: String, id: String })
-
-// prepare template data for display
-const icon = computed(() => getTypeIcon(props.type))
-const typeName = computed(() => getTypeInFrench(element.value?.newType || props.type))
-
-const franceAuthorization = computed(() => {
-  return element.value?.authorizationMode === "FR"
-})
-const authorizationInfo = computed(() => {
-  return {
-    flag: franceAuthorization.value ? "🇫🇷" : "🇪🇺",
-    text: franceAuthorization.value ? "Autorisé en France." : "Autorisé dans un état membre de l’EU ou EEE.",
-  }
-})
-
-const detailForType = {
-  plant: [
-    { label: "Nom", key: "newName" },
-    { label: "Description", key: "newDescription" },
-  ],
-  microorganism: [
-    { label: "Genre", key: "newGenre" },
-    { label: "Espèce", key: "newSpecies" },
-    { label: "Description", key: "newDescription" },
-  ],
-  default: [
-    { label: "Libillé", key: "newName" },
-    { label: "Description", key: "newDescription" },
-  ],
-}
-
-const elementProfile = computed(() => {
-  if (!element.value) return []
-
-  const items = []
-
-  const detail = detailForType[props.type] || detailForType.default
-  detail.forEach((d) => {
-    items.push({ label: d.label, text: element.value[d.key] })
-  })
-
-  if (franceAuthorization.value === false) {
-    items.push(
-      ...[
-        {
-          label: "Pays de référence",
-          text: element.value.euReferenceCountry,
-        },
-        {
-          label: "Source réglementaire",
-          text: element.value.euLegalSource,
-          href: element.value.euLegalSource,
-        },
-      ]
-    )
-  }
-  return items
-})
 
 const declarationId = computed(() => element.value?.declaration)
 const declarationLink = computed(() => {
